requirements-dev.txt
flask<3
Flask-Caching>=2.1.0
prometheus_client==0.19.0
json-logging==1.3.0
requests==2.31.0
Flask-Babel<3
werkzeug<2.3
black==23.12.1
py-healthcheck
pytest==7.4.4
pytest-testinfra==8.1.0
